lrucache: add func GetLocked(key string, f func (*[]byte)) bool
Avoid a copy by calling a supplied callback with a pointer to the data

A supplied callback function is called with the []byte representation
of a cached response only if it is found within the cache.
A bool will be returned set to true if the key was found, false otherwise
The supplied callback function is called while the lock is held.
The mutex is unlocked after the supplied function returns.

Signed-off-by: Michael Ira Krufky <mkrufky@gmail.com>